dvc: move Dvc/RepoTree to dvc/tree
This is where it really belongs.
